Move planner state to Zustand store (#89)
* Move planner state to Zustand store

* Restore DIN pallet orientation for wagon capacity

* Fix reset to clear derived planner state

* Fix reset to clear inputs and keep empty truck layout

* Fix pallet selection to continue after DIN saturation

* Improve euro auto layout selection

* Add per-entry stackable pallet selection
