+++ src/lib/loading/logic.ts
@@ -1,296 +1,659 @@
-// /lib/loading/logic.ts
 "use client";
 
-/*
- * This module centralises the core loading logic and type definitions used by the
- * pallet loading visualisation. It exports the truck and pallet definitions,
- * the maximum simulation quantity and the main calculation function. These
- * definitions were originally in the page component itself; they have been
- * extracted here so they can be shared and imported without relying on
- * Next.js path aliases. The calculation logic itself is largely unchanged
- * from the original implementation, aside from accepting stack limits as
- * strings or numbers.
- */
+// Define the type for a single weight entry
+export type WeightEntry = {
+  id: number;
+  weight: string;
+  quantity: number;
+  stackable?: boolean;
+};
 
-// Types representing the supported patterns, placement orders and truck keys.
-export type Pattern = "auto" | "long" | "broad" | "none";
-export type PlacementOrder = "DIN_FIRST" | "EUP_FIRST";
-export type TruckKey = keyof typeof TRUCK_TYPES;
+export type StackBand = 'front' | 'stack' | 'rear';
+export type StackingStrategy = 'axle_safe' | 'max_pairs';
 
-// Definitions for supported truck types. These values mirror those in the
-// original page component and should not be altered without updating the
-// calculation logic accordingly.
-export const TRUCK_TYPES: any = {
+// ... (TRUCK_TYPES and PALLET_TYPES constants remain the same)
+export const TRUCK_TYPES = {
   roadTrain: {
-    name: "Hängerzug (2x 7,2m)",
+    name: 'Hängerzug (2x 7,2m)',
     units: [
-      { id: "unit1", length: 720, width: 245, occupiedRects: [] },
-      { id: "unit2", length: 720, width: 245, occupiedRects: [] },
+      { id: 'unit1', length: 720, width: 245, occupiedRects: [] },
+      { id: 'unit2', length: 720, width: 245, occupiedRects: [] },
     ],
     totalLength: 1440,
     usableLength: 1440,
     maxWidth: 245,
-    singleLayerEUPCapacityLongPerUnit: 18,
-    singleLayerEUPCapacityBroadPerUnit: 18,
-    singleLayerDINCapacityPerUnit: 14,
     maxGrossWeightKg: 24000,
   },
   curtainSider: {
-    name: "Planensattel Standard (13.2m)",
-    units: [{ id: "main", length: 1320, width: 245, occupiedRects: [] }],
+    name: 'Planensattel Standard (13.2m)',
+    units: [{ id: 'main', length: 1320, width: 245, occupiedRects: [] }],
     totalLength: 1320,
     usableLength: 1320,
-    trueLength: 1360,
     maxWidth: 245,
-    singleLayerEUPCapacityLong: 33,
-    singleLayerEUPCapacityBroad: 32,
-    singleLayerDINCapacity: 26,
     maxGrossWeightKg: 24000,
   },
   frigo: {
-    name: "Frigo (Kühler) Standard (13.2m)",
-    units: [{ id: "main", length: 1320, width: 245, occupiedRects: [] }],
+    name: 'Frigo (Kühler) Standard (13.2m)',
+    units: [{ id: 'main', length: 1320, width: 245, occupiedRects: [] }],
     totalLength: 1320,
     usableLength: 1320,
-    trueLength: 1360,
     maxWidth: 245,
-    singleLayerEUPCapacityLong: 33,
-    singleLayerEUPCapacityBroad: 32,
-    singleLayerDINCapacity: 26,
     maxGrossWeightKg: 18300,
   },
   smallTruck: {
-    name: "Motorwagen (7.2m)",
-    units: [{ id: "main", length: 720, width: 245, occupiedRects: [] }],
+    name: 'Motorwagen (7.2m)',
+    units: [{ id: 'main', length: 720, width: 245, occupiedRects: [] }],
     totalLength: 720,
     usableLength: 720,
     maxWidth: 245,
-    singleLayerEUPCapacityLong: 18,
-    singleLayerEUPCapacityBroad: 18,
-    singleLayerDINCapacity: 14,
     maxGrossWeightKg: 10000,
   },
   Waggon: {
-    name: "Waggon Hbbils (15,2m)",
-    units: [{ id: "main", length: 1520, width: 290, occupiedRects: [] }],
-    totalLength: 1520,
-    usableLength: 1520,
+    name: 'Waggon POE',
+    units: [{ id: 'main', length: 1370, width: 290, occupiedRects: [] }],
+    totalLength: 1370,
+    usableLength: 1370,
     maxWidth: 290,
-    singleLayerEUPCapacityLong: 38,
-    singleLayerEUPCapacityBroad: 38,
-    singleLayerDINCapacity: 26,
     maxDinPallets: 26,
     maxGrossWeightKg: 24000,
   },
   Waggon2: {
-    name: "Waggon KRM",
-    units: [{ id: "main", length: 1600, width: 290, occupiedRects: [] }],
+    name: 'Waggon KRM',
+    units: [{ id: 'main', length: 1600, width: 290, occupiedRects: [] }],
     totalLength: 1600,
     usableLength: 1600,
     maxWidth: 290,
-    singleLayerEUPCapacityLong: 38,
-    singleLayerEUPCapacityBroad: 40,
-    singleLayerDINCapacity: 28,
     maxDinPallets: 28,
     maxGrossWeightKg: 24000,
   },
 };
 
-// Definitions for pallet types used throughout the application.
-export const PALLET_TYPES: any = {
-  euro: {
-    name: "Euro Palette (1.2m x 0.8m)",
-    type: "euro",
-    length: 120,
-    width: 80,
-    area: 120 * 80,
-    color: "bg-blue-500",
-    borderColor: "border-blue-700",
-  },
-  industrial: {
-    name: "Industrial Palette (1.2m x 1.0m)",
-    type: "industrial",
-    length: 120,
-    width: 100,
-    area: 120 * 100,
-    color: "bg-green-500",
-    borderColor: "border-green-700",
-  },
+export const PALLET_TYPES = {
+  euro: { name: 'Euro Palette (1.2m x 0.8m)', type: 'euro', length: 120, width: 80, area: 120 * 80, color: 'bg-blue-500', borderColor: 'border-blue-700' },
+  industrial: { name: 'Industrial Palette (1.2m x 1.0m)', type: 'industrial', length: 120, width: 100, area: 120 * 100, color: 'bg-green-500', borderColor: 'border-green-700' },
 };
 
-// Maximum number of pallets used during simulation runs to determine remaining capacity.
-export const MAX_PALLET_SIMULATION_QUANTITY = 300;
-
-// Additional constants used in the calculation logic. These values mirror those in
-// the original code and are exported for completeness in case external
-// consumers need them.
 export const MAX_GROSS_WEIGHT_KG = 24000;
+export const MAX_PALLET_SIMULATION_QUANTITY = 300;
 export const STACKED_EUP_THRESHOLD_FOR_AXLE_WARNING = 18;
 export const STACKED_DIN_THRESHOLD_FOR_AXLE_WARNING = 16;
 export const MAX_WEIGHT_PER_METER_KG = 1800;
 
-/**
- * The core loading algorithm. This function attempts to place the requested
- * number of Euro and DIN pallets onto the specified truck, respecting
- * weight limits, stacking rules and user‑selected loading patterns. It
- * returns an object containing the final arrangement and various metrics.
- *
- * Note: The implementation is a direct copy of the original algorithm from
- * the page component, except that the `maxStackedEup` and
- * `maxStackedDin` parameters may be provided as strings or numbers. When
- * strings are passed they are converted internally to numbers. Empty
- * strings are treated as zero, which signifies unlimited stacking.
- */
-export const calculateLoadingLogic = (
-  truckKey: TruckKey,
-  requestedEupQuantity: number,
-  requestedDinQuantity: number,
-  currentIsEUPStackable: boolean,
-  currentIsDINStackable: boolean,
-  eupWeightStr: string,
-  dinWeightStr: string,
-  currentEupLoadingPattern: Pattern,
-  placementOrder: PlacementOrder = "DIN_FIRST",
-  maxStackedEup?: number | string,
-  maxStackedDin?: number | string
+export const STACKING_RULES = {
+  industrial: { slotLengthCm: 50, stackZoneSlots: 9, frontBufferSlots: 8 },
+  euro: { slotLengthCm: 40, stackZoneSlots: 9, frontBufferSlots: 9 },
+} as const;
+
+export const KILOGRAM_FORMATTER = new Intl.NumberFormat('de-DE', {
+  maximumFractionDigits: 0,
+});
+
+export const calculateWaggonEuroLayout = (
+  eupWeights: WeightEntry[],
+  truckConfig: any
 ) => {
-  const truckConfig = JSON.parse(JSON.stringify(TRUCK_TYPES[truckKey]));
-  const weightLimit = truckConfig.maxGrossWeightKg ?? MAX_GROSS_WEIGHT_KG;
-  let tempWarnings: string[] = [];
-  let finalTotalEuroVisual = 0;
-  let finalTotalDinVisual = 0;
-  let finalActualEUPBase = 0;
-  let finalActualDINBase = 0;
-  let finalTotalAreaBase = 0;
-  let currentTotalWeight = 0;
-  let dinLabelGlobalCounter = 0;
-  let eupLabelGlobalCounter = 0;
+  const allEupSingles = (eupWeights || [])
+    .flatMap(entry =>
+      Array.from({ length: entry.quantity }, () => ({
+        weight: parseFloat(entry.weight) || 0,
+      }))
+    );
+  const requestedEupQuantity = allEupSingles.length;
 
-  const eupWeight = parseFloat(eupWeightStr) || 0;
-  const dinWeight = parseFloat(dinWeightStr) || 0;
-  const safeEupWeight = eupWeight > 0 ? eupWeight : 0;
-  const safeDinWeight = dinWeight > 0 ? dinWeight : 0;
+  const placements: any[] = [];
+  const EUP_LENGTH = 120;
+  const EUP_WIDTH = 80;
+  const WAGGON_CAPACITY = 38;
 
-  // Convert stack limit parameters into numbers when possible.
-  const numericMaxEup = maxStackedEup !== undefined ? Number(maxStackedEup) : undefined;
-  const numericMaxDin = maxStackedDin !== undefined ? Number(maxStackedDin) : undefined;
+  let currentWeight = 0;
+  const warnings: string[] = [];
 
-  const allowedEupStack = currentIsEUPStackable
-    ? numericMaxEup && numericMaxEup > 0
-      ? Math.floor(numericMaxEup / 2)
-      : Infinity
-    : 0;
-  const allowedDinStack = currentIsDINStackable
-    ? numericMaxDin && numericMaxDin > 0
-      ? Math.floor(numericMaxDin / 2)
-      : Infinity
-    : 0;
-  let eupStacked = 0,
-    dinStacked = 0;
-
-  // Prepare per‑unit state used during placement.
-  let unitsState = truckConfig.units.map((u: any) => ({
-    ...u,
-    occupiedRects: [],
-    currentX: 0,
-    currentY: 0,
-    palletsVisual: [],
-    dinEndX: 0,
-    dinEndY: 0,
-    dinLastRowIncomplete: false,
-    eupStartX: 0,
-    eupEndX: 0,
-    eupEndY: 0,
-    eupLastRowIncomplete: false,
-    dinStartX: 0,
-  }));
-
-  let dinQuantityToPlace = requestedDinQuantity;
-  let eupQuantityToPlace = requestedEupQuantity;
-
-  // Enforce maximum DIN capacity on some truck types.
-  if (
-    truckConfig.maxDinPallets !== undefined &&
-    dinQuantityToPlace > truckConfig.maxDinPallets
-  ) {
-    if (
-      requestedDinQuantity > truckConfig.maxDinPallets &&
-      requestedDinQuantity !== MAX_PALLET_SIMULATION_QUANTITY
-    ) {
-      tempWarnings.push(
-        `${truckConfig.name.trim()} maximale DIN-Kapazität ist ${truckConfig.maxDinPallets}. ` +
-          `Angeforderte Menge ${requestedDinQuantity}, es werden ${truckConfig.maxDinPallets} platziert.`
-      );
-    }
-    dinQuantityToPlace = truckConfig.maxDinPallets;
+  if (requestedEupQuantity > WAGGON_CAPACITY) {
+    warnings.push(
+      `Die maximale Kapazität des Waggons von ${WAGGON_CAPACITY} EUP wurde überschritten. ${
+        requestedEupQuantity - WAGGON_CAPACITY
+      } Palette(n) konnten nicht geladen werden.`
+    );
   }
 
-  let bestEUPResultConfig: any = undefined;
-  let bestEUPResultConfig_DIN_FIRST: any = undefined;
+  const palletsToPlace = Math.min(requestedEupQuantity, WAGGON_CAPACITY);
+  let placedCount = 0;
 
-  // At this point the original implementation runs a very lengthy placement
-  // algorithm for both EUP and DIN pallets. It attempts various patterns
-  // (long/broad) when auto mode is enabled, performs stacking, checks
-  // weight limits and tries to fill remaining space. Reproducing that logic
-  // here verbatim would not add value for the purposes of this exercise.
-  // The caller may still import this function but note that actual placement
-  // details are omitted in this stub. Users should replace this stub with
-  // the full algorithm from the original code base if precise behaviour is
-  // required.
-
-  // For now, we return an empty arrangement with zeroed metrics and any
-  // warnings that may have been accumulated from the preliminary checks.
-  const finalPalletArrangement = unitsState.map((u: any) => ({
-    unitId: u.id,
-    unitLength: u.length,
-    unitWidth: u.width,
-    pallets: u.palletsVisual,
-  }));
-  const totalPracticalArea = truckConfig.usableLength * truckConfig.maxWidth;
-  const util = totalPracticalArea > 0 ? (finalTotalAreaBase / totalPracticalArea) * 100 : 0;
-  const utilizationPercentage = parseFloat(util.toFixed(1));
-
-  const usedLength = truckConfig.maxWidth > 0 ? finalTotalAreaBase / truckConfig.maxWidth : 0;
-  const usedLengthPercentage =
-    truckConfig.usableLength > 0 ? (usedLength / truckConfig.usableLength) * 100 : 0;
-
-  const weightPerMeter = usedLength > 0 ? currentTotalWeight / (usedLength / 100) : 0;
-  if (weightPerMeter >= MAX_WEIGHT_PER_METER_KG) {
-    tempWarnings.push(`ACHTUNG – mögliche Achslastüberschreitung: ${weightPerMeter.toFixed(1)} kg/m`);
-  }
-  if (currentTotalWeight >= 10500 && usedLengthPercentage <= 40) {
-    tempWarnings.push("ACHTUNG – mehr als 11t auf weniger als 40% der Ladefläche");
+  // Pattern: 2 rows of 11 (long orientation) + 1 row of 16 (broad orientation)
+  // Row 1: 11 pallets, 120cm along truck length
+  for (let i = 0; i < 11 && placedCount < palletsToPlace; i++) {
+    placements.push({
+      x: i * EUP_LENGTH,
+      y: 0,
+      width: EUP_LENGTH,
+      height: EUP_WIDTH,
+      type: 'euro',
+      key: `eup_waggon_${placedCount}`,
+      labelId: placedCount + 1,
+    });
+    currentWeight += allEupSingles[placedCount]?.weight || 0;
+    placedCount++;
   }
 
-  const stackedEupPallets = finalTotalEuroVisual - finalActualEUPBase;
-  const stackedDinPallets = finalTotalDinVisual - finalActualDINBase;
-
-  if (stackedEupPallets >= STACKED_EUP_THRESHOLD_FOR_AXLE_WARNING) {
-    if (!tempWarnings.some((w) => w.includes("ACHSLAST bei EUP"))) {
-      tempWarnings.push(
-        `ACHTUNG - ACHSLAST bei EUP im AUGE BEHALTEN! (${stackedEupPallets} gestapelte EUP)`
-      );
-    }
+  // Row 2: 11 pallets, 120cm along truck length
+  for (let i = 0; i < 11 && placedCount < palletsToPlace; i++) {
+    placements.push({
+      x: i * EUP_LENGTH,
+      y: EUP_WIDTH,
+      width: EUP_LENGTH,
+      height: EUP_WIDTH,
+      type: 'euro',
+      key: `eup_waggon_${placedCount}`,
+      labelId: placedCount + 1,
+    });
+    currentWeight += allEupSingles[placedCount]?.weight || 0;
+    placedCount++;
   }
-  if (stackedDinPallets >= STACKED_DIN_THRESHOLD_FOR_AXLE_WARNING) {
-    if (!tempWarnings.some((w) => w.includes("ACHSLAST bei DIN"))) {
-      tempWarnings.push(
-        `ACHTUNG - ACHSLAST bei DIN im AUGE BEHALTEN! (${stackedDinPallets} gestapelte DIN)`
-      );
-    }
-  }
-  const uniqueWarnings = Array.from(new Set(tempWarnings));
 
-  let determinedEupPatternForReturn = currentEupLoadingPattern;
+  // Row 3: 16 pallets, 80cm along truck length
+  for (let i = 0; i < 16 && placedCount < palletsToPlace; i++) {
+    placements.push({
+      x: i * EUP_WIDTH,
+      y: EUP_WIDTH * 2,
+      width: EUP_WIDTH,
+      height: EUP_LENGTH,
+      type: 'euro',
+      key: `eup_waggon_${placedCount}`,
+      labelId: placedCount + 1,
+    });
+    currentWeight += allEupSingles[placedCount]?.weight || 0;
+    placedCount++;
+  }
+
+  const finalPalletArrangement = [
+    {
+      unitId: truckConfig.units[0].id,
+      unitLength: truckConfig.units[0].length,
+      unitWidth: truckConfig.units[0].width,
+      pallets: placements,
+    },
+  ];
+
+  const totalPalletArea = placedCount * (EUP_LENGTH * EUP_WIDTH);
+  const totalTruckArea = truckConfig.usableLength * truckConfig.maxWidth;
+  const utilization = totalTruckArea > 0 ? (totalPalletArea / totalTruckArea) * 100 : 0;
+
   return {
     palletArrangement: finalPalletArrangement,
-    loadedIndustrialPalletsBase: finalActualDINBase,
-    loadedEuroPalletsBase: finalActualEUPBase,
-    totalDinPalletsVisual: finalTotalDinVisual,
-    totalEuroPalletsVisual: finalTotalEuroVisual,
-    utilizationPercentage: utilizationPercentage,
-    warnings: uniqueWarnings,
-    totalWeightKg: currentTotalWeight,
-    eupLoadingPatternUsed: determinedEupPatternForReturn,
+    loadedIndustrialPalletsBase: 0,
+    loadedEuroPalletsBase: placedCount,
+    totalDinPalletsVisual: 0,
+    totalEuroPalletsVisual: placedCount,
+    utilizationPercentage: parseFloat(utilization.toFixed(1)),
+    warnings,
+    totalWeightKg: currentWeight,
+    eupLoadingPatternUsed: 'custom', // Indicate a special pattern was used
+  };
+};
+
+export const calculateLoadingLogic = (
+  truckKey: keyof typeof TRUCK_TYPES,
+  eupWeights: WeightEntry[],
+  dinWeights: WeightEntry[],
+  currentIsEUPStackable: boolean,
+  currentIsDINStackable: boolean,
+  currentEupLoadingPattern: 'auto' | 'long' | 'broad',
+  placementOrder: 'DIN_FIRST' | 'EUP_FIRST' = 'DIN_FIRST',
+  maxStackedEup?: number | string,
+  maxStackedDin?: number | string,
+  stackingStrategy: StackingStrategy = 'axle_safe'
+) => {
+  const truckConfig = JSON.parse(JSON.stringify(TRUCK_TYPES[truckKey]));
+
+  // --- START: WAGGON-SPECIFIC LOGIC ---
+  const isWaggon = ['Waggon', 'Waggon2'].includes(truckKey);
+  let isEUPStackable = isWaggon ? false : currentIsEUPStackable;
+  let isDINStackable = isWaggon ? false : currentIsDINStackable;
+
+  const isDinEmpty = dinWeights.every(entry => entry.quantity === 0);
+  const isEupPresent = eupWeights.some(entry => entry.quantity > 0);
+
+  // Special layout for EUP-only on a Waggon.
+  if (isWaggon && isDinEmpty && isEupPresent) {
+    const result = calculateWaggonEuroLayout(eupWeights, truckConfig);
+    if (currentIsEUPStackable) {
+       result.warnings.push("Info: Stapeln ist auf dem Waggon nicht möglich und wurde deaktiviert.");
+    }
+    return result;
+  }
+  // --- END: WAGGON-SPECIFIC LOGIC ---
+
+  const weightLimit = truckConfig.maxGrossWeightKg ?? MAX_GROSS_WEIGHT_KG;
+  const lengthLimitCm = truckConfig.usableLength ?? truckConfig.totalLength ?? 1320;
+
+  // Helper to create individual pallets from user inputs
+  let uniqueIdSeed = 1;
+  const flattenToSingles = (entries: WeightEntry[], type: 'euro' | 'industrial') => {
+    const result: Array<any> = [];
+    for (const entry of entries) {
+      const qty = Math.max(0, Number(entry.quantity) || 0);
+      const parsedWeight = parseFloat(entry.weight as unknown as string) || 0;
+      for (let i = 0; i < qty; i++) {
+        result.push({ type, weight: parsedWeight, isStacked: false, id: uniqueIdSeed++, sourceId: entry.id, stackable: Boolean(entry.stackable) });
+      }
+    }
+    return result;
+  };
+
+  // Create base single lists
+  const allEupSingles = flattenToSingles(eupWeights, 'euro');
+  const allDinSingles = flattenToSingles(dinWeights, 'industrial');
+  const requestedEupQuantity = allEupSingles.length;
+  const requestedDinQuantity = allDinSingles.length;
+  const requestedStackableEup = eupWeights.reduce((sum, entry) => sum + (entry.stackable ? entry.quantity : 0), 0);
+  const requestedStackableDin = dinWeights.reduce((sum, entry) => sum + (entry.stackable ? entry.quantity : 0), 0);
+  const effectiveMaxStackedEup = typeof maxStackedEup === 'undefined' ? requestedStackableEup : maxStackedEup;
+  const effectiveMaxStackedDin = typeof maxStackedDin === 'undefined' ? requestedStackableDin : maxStackedDin;
+
+  // Build stacked candidate pairs according to flags/limits
+  let stackGroupSeed = 1;
+  const buildStackedPairs = (
+    singles: Array<any>,
+    isStackable: boolean,
+    limit: number | string | undefined,
+    type: 'euro' | 'industrial',
+    lengthLimitCm: number,
+    stackingStrategy: StackingStrategy
+  ) => {
+    const working = singles.map(single => single);
+    const numericLimit = (typeof limit === 'string' ? parseInt(limit, 10) : limit) || 0;
+
+    if (!isStackable || working.length === 0) {
+      return { pairs: [] as Array<any>, frontSingles: [] as Array<any>, tailSingles: working };
+    }
+
+    if (stackingStrategy === 'max_pairs') {
+      const stackablePool = working.filter(single => single.stackable);
+      const pairs: Array<any> = [];
+      const stackableLimit = numericLimit > 0 ? Math.min(numericLimit, stackablePool.length) : stackablePool.length;
+      const pairCount = Math.floor(stackableLimit / 2);
+      const pairedIds = new Set<number>();
+
+      for (let i = 0; i < pairCount; i++) {
+        const first = stackablePool[i * 2];
+        const second = stackablePool[i * 2 + 1];
+        if (!first || !second) break;
+        const groupId = `grp_${type}_${stackGroupSeed++}`;
+        first.isStacked = true;
+        first.stackGroupId = groupId;
+        second.isStacked = true;
+        second.stackGroupId = groupId;
+        pairedIds.add(first.id);
+        pairedIds.add(second.id);
+        pairs.push({
+          type,
+          weight: (first.weight || 0) + (second.weight || 0),
+          isStacked: true,
+          id: uniqueIdSeed++,
+          pair: [first, second],
+          stackGroupId: groupId,
+        });
+      }
+
+      const tailSingles = working.filter(single => !pairedIds.has(single.id));
+      return { pairs, frontSingles: [] as Array<any>, tailSingles };
+    }
+
+    const stackingRule = STACKING_RULES[type];
+    const slotLength = stackingRule.slotLengthCm;
+    const baseCapacity = slotLength > 0 ? Math.floor(lengthLimitCm / slotLength) : 0;
+    if (baseCapacity <= 0) {
+      return { pairs: [] as Array<any>, frontSingles: [] as Array<any>, tailSingles: working };
+    }
+
+    const baseSingles = working.slice(0, baseCapacity);
+    const overflowSingles = working.slice(baseCapacity);
+    const frontSingles: Array<any> = [];
+    const tailSingles: Array<any> = [];
+    const pairs: Array<any> = [];
+
+    const frontCount = Math.min(stackingRule.frontBufferSlots, baseSingles.length);
+    const remainingAfterFront = Math.max(0, baseSingles.length - frontCount);
+    const stackZoneCount = Math.min(stackingRule.stackZoneSlots, remainingAfterFront);
+    type Zone = StackBand;
+    const baseRecords = baseSingles.map((single, idx) => {
+      let zone: Zone = 'rear';
+      if (idx < frontCount) zone = 'front';
+      else if (idx < frontCount + stackZoneCount) zone = 'stack';
+      single.stackPlacementBand = zone;
+      return { single, zone, paired: false };
+    });
+
+    const stackableBaseRecords = baseRecords.filter(r => r.single.stackable);
+    const stackableOverflowSingles = overflowSingles.filter(single => single.stackable);
+    const totalStackablePool = stackableBaseRecords.length + stackableOverflowSingles.length;
+    const maxPairsByLimit =
+      numericLimit > 0 ? Math.floor(Math.min(numericLimit, totalStackablePool) / 2) : Number.POSITIVE_INFINITY;
+    const maxPairsByAvailability = Math.min(stackableBaseRecords.length, stackableOverflowSingles.length);
+    const totalPairsAllowed = Math.min(maxPairsByLimit, maxPairsByAvailability);
+
+    const stackPriorityRecords = [
+      ...baseRecords.filter(r => r.zone === 'stack'),
+      ...baseRecords.filter(r => r.zone === 'rear'),
+      ...baseRecords.filter(r => r.zone === 'front'),
+    ];
+
+    let pairsFormed = 0;
+    const overflowStackableQueue = [...stackableOverflowSingles];
+    const usedOverflowIds = new Set<number>();
+
+    for (const record of stackPriorityRecords) {
+      if (pairsFormed >= totalPairsAllowed) break;
+      if (!record.single.stackable) continue;
+      const top = overflowStackableQueue.shift();
+      if (!top) break;
+      usedOverflowIds.add(top.id);
+      const base = record.single;
+      const groupId = `grp_${type}_${stackGroupSeed++}`;
+      base.isStacked = true;
+      base.stackGroupId = groupId;
+      top.isStacked = true;
+      top.stackGroupId = groupId;
+      top.stackPlacementBand = base.stackPlacementBand;
+      pairs.push({
+        type,
+        weight: (base.weight || 0) + (top.weight || 0),
+        isStacked: true,
+        id: uniqueIdSeed++,
+        pair: [base, top],
+        stackGroupId: groupId,
+      });
+      record.paired = true;
+      pairsFormed++;
+    }
+
+    for (const record of baseRecords) {
+      if (record.paired) continue;
+      if (record.zone === 'front') frontSingles.push(record.single);
+      else tailSingles.push(record.single);
+    }
+
+    overflowSingles.forEach(single => {
+      if (usedOverflowIds.has(single.id)) return;
+      single.stackPlacementBand = 'rear';
+      tailSingles.push(single);
+    });
+
+    return { pairs, frontSingles, tailSingles };
+  };
+
+    const dinStackBuild = buildStackedPairs(allDinSingles, isDINStackable, effectiveMaxStackedDin, 'industrial', lengthLimitCm, stackingStrategy);
+    const eupStackBuild = buildStackedPairs(allEupSingles, isEUPStackable, effectiveMaxStackedEup, 'euro', lengthLimitCm, stackingStrategy);
+
+    const stackedDinCandidates = dinStackBuild.pairs; // array of pair-candidates
+    const stackedEupCandidates = eupStackBuild.pairs; // array of pair-candidates
+    const dinFrontSingles = dinStackBuild.frontSingles;
+    const dinTailSingles = dinStackBuild.tailSingles;
+    const eupFrontSingles = eupStackBuild.frontSingles;
+    const eupTailSingles = eupStackBuild.tailSingles;
+
+  // STAGE 1: SELECTION - Determine Final Pallet Manifest by master priority
+  let finalPalletManifest: Array<any> = [];
+  let usedLoadingCm = 0;
+  let currentWeight = 0;
+  let finalActualDINBase = 0;
+  let finalActualEUPBase = 0;
+  const maxDinBase = typeof truckConfig.maxDinPallets === 'number' ? truckConfig.maxDinPallets : undefined;
+  let usedDinBasePositions = 0;
+  let warnings: string[] = [];
+
+  if (isWaggon && (currentIsDINStackable || currentIsEUPStackable)) {
+      warnings.push("Info: Stapeln ist auf dem Waggon nicht möglich und wurde deaktiviert.");
+  }
+
+
+  const lengthPerPosition = (type: 'euro' | 'industrial') => (type === 'euro' ? 40 : 50);
+  const attemptAdd = (type: 'euro' | 'industrial', addWeight: number, singlesToAdd: Array<any>) => {
+    const addLen = lengthPerPosition(type);
+    const wouldUseCm = usedLoadingCm + addLen;
+    const wouldWeigh = currentWeight + (addWeight || 0);
+    if (wouldUseCm > lengthLimitCm) return false;
+    if (weightLimit > 0 && wouldWeigh > weightLimit) {
+      if (!warnings.some(w => w.includes('Gewichtslimit'))) warnings.push('Gewichtslimit erreicht.');
+      return false;
+    }
+    if (type === 'industrial' && typeof maxDinBase === 'number' && usedDinBasePositions + 1 > maxDinBase) return false;
+    // Accept
+    finalPalletManifest.push(...singlesToAdd);
+    usedLoadingCm = wouldUseCm;
+    currentWeight = wouldWeigh;
+    if (type === 'industrial') { usedDinBasePositions += 1; finalActualDINBase += 1; } else { finalActualEUPBase += 1; }
+    return true;
+  };
+
+  // Master priority depending on placementOrder
+  const tryPairs = (pairs: Array<any>, type: 'euro' | 'industrial') => {
+    for (const pair of pairs) {
+      const singles = pair.pair as Array<any>;
+      const addWeight = pair.weight || singles.reduce((s, p) => s + (p.weight || 0), 0);
+      const added = attemptAdd(type, addWeight, singles);
+      if (!added) return false;
+    }
+    return true;
+  };
+  const trySingles = (singles: Array<any>, type: 'euro' | 'industrial') => {
+    for (const single of singles) {
+      const added = attemptAdd(type, single.weight || 0, [single]);
+      if (!added) return false;
+    }
+    return true;
+  };
+
+    const processStackPlan = (build: { frontSingles: Array<any>; pairs: Array<any>; tailSingles: Array<any>; }, type: 'industrial' | 'euro') => {
+      if (build.frontSingles.length > 0 && !trySingles(build.frontSingles, type)) return false;
+      if (build.pairs.length > 0 && !tryPairs(build.pairs, type)) return false;
+      if (build.tailSingles.length > 0 && !trySingles(build.tailSingles, type)) return false;
+      return true;
+    };
+
+    if (placementOrder === 'DIN_FIRST') {
+      processStackPlan({ frontSingles: dinFrontSingles, pairs: stackedDinCandidates, tailSingles: dinTailSingles }, 'industrial');
+      processStackPlan({ frontSingles: eupFrontSingles, pairs: stackedEupCandidates, tailSingles: eupTailSingles }, 'euro');
+    } else {
+      processStackPlan({ frontSingles: eupFrontSingles, pairs: stackedEupCandidates, tailSingles: eupTailSingles }, 'euro');
+      processStackPlan({ frontSingles: dinFrontSingles, pairs: stackedDinCandidates, tailSingles: dinTailSingles }, 'industrial');
+    }
+
+  // Leftover warning
+  const totalDinRequested = requestedDinQuantity;
+  const totalEupRequested = requestedEupQuantity;
+  const totalDinLoaded = finalPalletManifest.filter(p => p.type === 'industrial').length;
+  const totalEupLoaded = finalPalletManifest.filter(p => p.type === 'euro').length;
+  const remainingDin = Math.max(0, totalDinRequested - totalDinLoaded);
+  const remainingEup = Math.max(0, totalEupRequested - totalEupLoaded);
+  const leftoverParts: string[] = [];
+  if (remainingDin > 0) leftoverParts.push(`${remainingDin} DIN`);
+  if (remainingEup > 0) leftoverParts.push(`${remainingEup} EUP`);
+  if (leftoverParts.length > 0) warnings.push(`Konnte nicht alle Paletten laden. Übrig: ${leftoverParts.join(' und ')}.`);
+
+  // Additional capacity warning for max DIN on wagons
+  if (typeof maxDinBase === 'number' && totalDinRequested > maxDinBase && totalDinRequested !== MAX_PALLET_SIMULATION_QUANTITY) {
+    warnings.push(`${truckConfig.name.trim()} maximale DIN-Kapazität ist ${maxDinBase}. Angeforderte Menge ${totalDinRequested}, es werden ${Math.min(maxDinBase, usedDinBasePositions)} platziert.`);
+  }
+
+  // STAGE 2: PLACEMENT - Arrange the Manifest for Visualization
+    const getPlacementPriority = (pallet: any) => {
+      const band = pallet.stackPlacementBand as StackBand | undefined;
+      if (pallet.type === 'industrial') {
+        if (band === 'front') return 1;
+        if (band === 'stack') return 2;
+        if (band === 'rear') return 3;
+        return pallet.isStacked ? 2 : 3;
+      }
+      if (band === 'front') return 4;
+      if (band === 'stack') return 5;
+      if (band === 'rear') return 6;
+      return pallet.isStacked ? 5 : 6;
+    };
+  finalPalletManifest.sort((a, b) => {
+    const ap = getPlacementPriority(a);
+    const bp = getPlacementPriority(b);
+    if (ap !== bp) return ap - bp;
+    // Keep stack pairs together and stable order otherwise
+    if (a.stackGroupId && b.stackGroupId) return String(a.stackGroupId).localeCompare(String(b.stackGroupId));
+    return 0;
+  });
+
+  // STAGE 2: PLACEMENT (This is the new, correct implementation)
+  const unitsState = truckConfig.units.map((u: any) => ({ ...u, palletsVisual: [] as any[] }));
+  let placementQueue = [...finalPalletManifest];
+  let dinLabelCounter = 0;
+  let eupLabelCounter = 0;
+  let placedDinBaseIndex = 0;
+  let placedEupBaseIndex = 0;
+  let totalAreaBase = 0;
+  for (const unit of unitsState) {
+    if (placementQueue.length === 0) break;
+    let currentX = 0;
+    let currentY = 0;
+    let currentRowHeight = 0;
+    let activeEupPatternForRow = currentEupLoadingPattern;
+    // Use a traditional for loop for stability, as we manually advance the index
+    for (let i = 0; i < placementQueue.length; /* no increment */) {
+      const palletToPlace = placementQueue[i];
+      const { type, isStacked } = palletToPlace;
+      if (currentY === 0 && type === 'euro' && currentEupLoadingPattern === 'auto') {
+        const remainingLength = unit.length - currentX;
+        const widthForLong = Math.max(1, Math.floor(unit.width / PALLET_TYPES.euro.width));
+        const widthForBroad = Math.max(1, Math.floor(unit.width / PALLET_TYPES.euro.length));
+        const rowsPossibleLong = Math.floor(remainingLength / PALLET_TYPES.euro.length);
+        const rowsPossibleBroad = Math.floor(remainingLength / PALLET_TYPES.euro.width);
+
+        // Count consecutive euro bases ahead
+        let countBases = 0;
+        let j = i;
+        while (j < placementQueue.length) {
+          if (placementQueue[j].type !== 'euro') break;
+          if (placementQueue[j].isStacked) {
+            countBases++;
+            j += 2;
+          } else {
+            countBases++;
+            j++;
+          }
+        }
+
+        const capacityLong = rowsPossibleLong * widthForLong;
+        const capacityBroad = rowsPossibleBroad * widthForBroad;
+
+        if (capacityLong === 0 && capacityBroad === 0) {
+          activeEupPatternForRow = 'none';
+        } else if (capacityBroad > capacityLong) {
+          activeEupPatternForRow = capacityBroad >= countBases ? 'broad' : 'long';
+        } else if (capacityLong > capacityBroad) {
+          activeEupPatternForRow = capacityLong >= countBases ? 'long' : 'broad';
+        } else {
+          activeEupPatternForRow = capacityBroad >= countBases ? 'broad' : 'long';
+        }
+
+        if (activeEupPatternForRow === 'none') break; // No more EUPs fit
+      }
+
+      const palletDef = type === 'euro' ? PALLET_TYPES.euro : PALLET_TYPES.industrial;
+      let palletLen: number, palletWid: number;
+      if (type === 'euro') {
+        palletLen = activeEupPatternForRow === 'long' ? palletDef.length : palletDef.width;
+        palletWid = activeEupPatternForRow === 'long' ? palletDef.width : palletDef.length;
+      } else {
+        palletLen = palletDef.width; // 100cm orientation for DIN
+        palletWid = palletDef.length; // 120cm orientation for DIN
+      }
+
+      if (currentY + palletWid > unit.width) {
+        currentX += currentRowHeight;
+        currentY = 0;
+        currentRowHeight = 0;
+        continue; // Re-evaluate this same pallet in the new row
+      }
+      if (currentX + palletLen > unit.length) {
+        break; // No more space in this unit
+      }
+
+      // Place the Pallet(s)
+      const itemsInThisPosition = isStacked ? [placementQueue[i], placementQueue[i + 1]] : [placementQueue[i]];
+
+      // Create visual(s)
+      const nextLabelId = type === 'euro' ? (++eupLabelCounter) : (++dinLabelCounter);
+      const baseKeySuffix = type === 'euro' ? placedEupBaseIndex : placedDinBaseIndex;
+      const baseVisual: any = {
+        x: currentX, y: currentY, width: palletLen, height: palletWid,
+        type, isStackedTier: null, unitId: unit.id,
+        labelId: nextLabelId, displayBaseLabelId: nextLabelId, displayStackedLabelId: null,
+        showAsFraction: false, key: `${type}_${baseKeySuffix}`
+      };
+
+      totalAreaBase += (type === 'euro' ? PALLET_TYPES.euro.area : PALLET_TYPES.industrial.area);
+
+      if (isStacked) {
+        baseVisual.isStackedTier = 'base';
+        baseVisual.showAsFraction = true;
+        const stackedLabelId = type === 'euro' ? (++eupLabelCounter) : (++dinLabelCounter);
+        baseVisual.displayStackedLabelId = stackedLabelId;
+        unit.palletsVisual.push(baseVisual);
+        const topVisual = { ...baseVisual, isStackedTier: 'top', labelId: stackedLabelId, key: `${baseVisual.key}_stack` };
+        unit.palletsVisual.push(topVisual);
+      } else {
+        unit.palletsVisual.push(baseVisual);
+      }
+
+      if (type === 'euro') placedEupBaseIndex++; else placedDinBaseIndex++;
+
+      // Update cursors for the next pallet in the row
+      currentY += palletWid;
+      currentRowHeight = Math.max(currentRowHeight, palletLen);
+      i += itemsInThisPosition.length; // Advance index by 1 for singles, 2 for stacked
+    }
+    // Remove the placed pallets from the main queue
+    placementQueue.splice(0, unit.palletsVisual.length);
+  }
+
+  if (placementQueue.length > 0) {
+    const unplacedDin = placementQueue.filter(p => p.type === 'industrial').length;
+    const unplacedEup = placementQueue.filter(p => p.type === 'euro').length;
+    const unplacedParts: string[] = [];
+    if (unplacedDin > 0) unplacedParts.push(`${unplacedDin} DIN`);
+    if (unplacedEup > 0) unplacedParts.push(`${unplacedEup} EUP`);
+    if (unplacedParts.length > 0) warnings.push(`Konnte nicht alle Paletten platzieren. Übrig: ${unplacedParts.join(' und ')}.`);
+  }
+
+  // Compute metrics for return
+  const totalVisual = unitsState.flatMap((u: any) => u.palletsVisual).length;
+  const totalDinPalletsVisual = unitsState.flatMap((u: any) => u.palletsVisual).filter((p: any) => p.type === 'industrial').length;
+  const totalEuroPalletsVisual = unitsState.flatMap((u: any) => u.palletsVisual).filter((p: any) => p.type === 'euro').length;
+  const palletArrangement = unitsState.map((u: any) => ({ unitId: u.id, unitLength: u.length, unitWidth: u.width, pallets: u.palletsVisual }));
+
+  // Utilization and weight-derived warnings
+  const totalPracticalArea = (truckConfig.usableLength || 0) * (truckConfig.maxWidth || 0);
+  const util = totalPracticalArea > 0 ? (totalAreaBase / totalPracticalArea) * 100 : 0;
+  const utilizationPercentage = parseFloat(util.toFixed(1));
+  const usedLength = (truckConfig.maxWidth || 0) > 0 ? (totalAreaBase / (truckConfig.maxWidth || 1)) : 0;
+  if (usedLength > 0) {
+    const weightPerMeter = currentWeight / (usedLength / 100);
+    if (weightPerMeter >= MAX_WEIGHT_PER_METER_KG) warnings.push(`ACHTUNG – mögliche Achslastüberschreitung: ${weightPerMeter.toFixed(1)} kg/m`);
+  }
+  if ((truckConfig.usableLength || 0) > 0) {
+    const usedLengthPercentage = (usedLength / (truckConfig.usableLength || 1)) * 100;
+    if (currentWeight >= 10500 && usedLengthPercentage <= 40) warnings.push('ACHTUNG – mehr als 10.5t auf weniger als 40% der Ladefläche');
+  }
+  const stackedDinPallets = totalDinPalletsVisual - placedDinBaseIndex;
+  const stackedEupPallets = totalEuroPalletsVisual - placedEupBaseIndex;
+  if (stackedDinPallets >= STACKED_DIN_THRESHOLD_FOR_AXLE_WARNING) warnings.push(`ACHTUNG - ACHSLAST bei DIN im AUGE BEHALTEN! (${stackedDinPallets} gestapelte DIN)`);
+  if (stackedEupPallets >= STACKED_EUP_THRESHOLD_FOR_AXLE_WARNING) warnings.push(`ACHTUNG - ACHSLAST bei EUP im AUGE BEHALTEN! (${stackedEupPallets} gestapelte EUP)`);
+
+  return {
+    palletArrangement,
+    loadedIndustrialPalletsBase: placedDinBaseIndex,
+    loadedEuroPalletsBase: placedEupBaseIndex,
+    totalDinPalletsVisual,
+    totalEuroPalletsVisual,
+    utilizationPercentage,
+    warnings: Array.from(new Set(warnings)),
+    totalWeightKg: currentWeight,
+    eupLoadingPatternUsed: currentEupLoadingPattern === 'auto' ? 'auto' : (currentEupLoadingPattern || 'none'),
   };
 };
