+++ src/store.ts
@@ -1,117 +1,375 @@
 import { create } from 'zustand';
 import { devtools } from 'zustand/middleware';
-import { CONTAINER_PRESETS, PALLET_PRESETS } from './presets';
-import type { Constraints, Plan, Units } from './types';
-import { computePlan } from './lib/layoutEngine';
-import { PlannerInputSchema, type PlannerInputs } from './validation';
+import type { PalletUnitArrangement } from './components/TruckVisualization';
+import {
+  MAX_PALLET_SIMULATION_QUANTITY,
+  TRUCK_TYPES,
+  calculateLoadingLogic,
+  type StackingStrategy,
+  type WeightEntry,
+} from './lib/loading/logic';
 
-interface PlannerState {
-  // inputs
-  units: Units;
-  containerId: string; // selected preset id or 'custom'
-  containerCustom?: PlannerInputs['container'];
-  palletId: string; // selected preset id or 'custom'
-  palletCustom?: PlannerInputs['pallet'];
-  constraints: Constraints;
-  note?: string;
+type LastEdited = 'eup' | 'din';
 
-  // derived
-  plan: Plan | null;
-  lastValidationErrors: string[]; // compact errors for non-UI logging/inspection
+type PlannerState = {
+  selectedTruck: keyof typeof TRUCK_TYPES;
+  eupWeights: WeightEntry[];
+  dinWeights: WeightEntry[];
+  eupLoadingPattern: 'auto' | 'long' | 'broad';
+  isEUPStackable: boolean;
+  isDINStackable: boolean;
+  stackingStrategy: StackingStrategy;
+  loadedEuroPalletsBase: number;
+  loadedIndustrialPalletsBase: number;
+  totalEuroPalletsVisual: number;
+  totalDinPalletsVisual: number;
+  utilizationPercentage: number;
+  warnings: string[];
+  palletArrangement: PalletUnitArrangement[];
+  totalWeightKg: number;
+  actualEupLoadingPattern: string;
+  remainingCapacity: { eup: number; din: number };
+  lastEdited: LastEdited;
 
-  // actions
-  setUnits: (units: Units) => void;
-  setContainerById: (id: string) => void;
-  setCustomContainer: (container: PlannerInputs['container']) => void;
-  setPalletById: (id: string) => void;
-  setCustomPallet: (pallet: PlannerInputs['pallet']) => void;
-  setConstraints: (constraints: Constraints) => void;
-  setNote: (note?: string) => void;
-  recompute: () => void;
-  reset: () => void;
-}
+  setSelectedTruck: (truck: keyof typeof TRUCK_TYPES) => void;
+  setEupWeights: (weights: WeightEntry[]) => void;
+  setDinWeights: (weights: WeightEntry[]) => void;
+  setEupLoadingPattern: (pattern: 'auto' | 'long' | 'broad') => void;
+  setIsEUPStackable: (stackable: boolean) => void;
+  setIsDINStackable: (stackable: boolean) => void;
+  setStackingStrategy: (strategy: StackingStrategy) => void;
+  setLastEdited: (last: LastEdited) => void;
 
-function getSelectedContainer(state: PlannerState): PlannerInputs['container'] {
-  if (state.containerId === 'custom' && state.containerCustom) return state.containerCustom;
-  const preset = CONTAINER_PRESETS.find((c) => c.id === state.containerId) ?? CONTAINER_PRESETS[0];
-  return preset;
-}
-
-function getSelectedPallet(state: PlannerState): PlannerInputs['pallet'] {
-  if (state.palletId === 'custom' && state.palletCustom) return state.palletCustom;
-  const preset = PALLET_PRESETS.find((p) => p.id === state.palletId) ?? PALLET_PRESETS[0];
-  return preset;
-}
-
-function computeFromState(state: PlannerState): { plan: Plan | null; errors: string[] } {
-  const container = getSelectedContainer(state);
-  const pallet = getSelectedPallet(state);
-
-  const candidate: PlannerInputs = {
-    container,
-    pallet,
-    constraints: state.constraints,
-    units: state.units,
-    note: state.note,
-  };
-
-  const parsed = PlannerInputSchema.safeParse(candidate);
-  if (!parsed.success) {
-    const errors = parsed.error.issues.map((i) => `${i.path.join('.')}: ${i.message}`);
-    return { plan: null, errors };
-  }
-
-  const plan = computePlan({
-    container: parsed.data.container,
-    pallet: parsed.data.pallet,
-    constraints: parsed.data.constraints,
-    units: parsed.data.units,
-    note: parsed.data.note,
-  });
-
-  return { plan, errors: [] };
-}
-
-const defaultConstraints: Constraints = {
-  allowRotate: true,
-  wallClearance: 60,
-  betweenClearance: 50,
-  aisleLengthReserve: 0,
+  clearAllPallets: () => void;
+  maximizePallets: (palletTypeToMax: 'euro' | 'industrial') => void;
+  fillRemaining: (typeToFill: 'euro' | 'industrial') => boolean;
+  recalculate: () => void;
 };
+
+const createEmptyWeightEntries = () => ({
+  eupWeights: [{ id: Date.now(), weight: '', quantity: 0, stackable: false }],
+  dinWeights: [{ id: Date.now() + 1, weight: '', quantity: 0, stackable: false }],
+});
+
+const computeStackableCount = (entries: WeightEntry[]) =>
+  entries.reduce((sum, entry) => sum + (entry.stackable ? entry.quantity : 0), 0);
 
 export const usePlannerStore = create<PlannerState>()(
   devtools((set, get) => ({
-    units: 'metric',
-    containerId: CONTAINER_PRESETS[0]?.id ?? 'sprinter',
-    palletId: PALLET_PRESETS[0]?.id ?? 'euro',
-    constraints: defaultConstraints,
-    note: undefined,
-    plan: null,
-    lastValidationErrors: [],
+    selectedTruck: 'curtainSider',
+    ...createEmptyWeightEntries(),
+    eupLoadingPattern: 'auto',
+    isEUPStackable: false,
+    isDINStackable: false,
+    stackingStrategy: 'axle_safe',
+    loadedEuroPalletsBase: 0,
+    loadedIndustrialPalletsBase: 0,
+    totalEuroPalletsVisual: 0,
+    totalDinPalletsVisual: 0,
+    utilizationPercentage: 0,
+    warnings: [],
+    palletArrangement: [],
+    totalWeightKg: 0,
+    actualEupLoadingPattern: 'auto',
+    remainingCapacity: { eup: 0, din: 0 },
+    lastEdited: 'eup',
 
-    setUnits: (units) => set({ units }, false, 'setUnits'),
-    setContainerById: (id) => set({ containerId: id, containerCustom: undefined }, false, 'setContainerById'),
-    setCustomContainer: (container) => set({ containerId: 'custom', containerCustom: container }, false, 'setCustomContainer'),
-    setPalletById: (id) => set({ palletId: id, palletCustom: undefined }, false, 'setPalletById'),
-    setCustomPallet: (pallet) => set({ palletId: 'custom', palletCustom: pallet }, false, 'setCustomPallet'),
-    setConstraints: (constraints) => set({ constraints }, false, 'setConstraints'),
-    setNote: (note) => set({ note }, false, 'setNote'),
+    setSelectedTruck: (truck) =>
+      set(() => {
+        const updates: Partial<PlannerState> = { selectedTruck: truck };
+        if (['Waggon', 'Waggon2'].includes(truck)) {
+          updates.isEUPStackable = false;
+          updates.isDINStackable = false;
+        }
+        return updates;
+      }, false, 'setSelectedTruck'),
 
-    recompute: () => {
-      const results = computeFromState(get());
-      set({ plan: results.plan, lastValidationErrors: results.errors }, false, 'recompute');
+    setEupWeights: (weights) => set({ eupWeights: weights }, false, 'setEupWeights'),
+    setDinWeights: (weights) => set({ dinWeights: weights }, false, 'setDinWeights'),
+    setEupLoadingPattern: (pattern) => set({ eupLoadingPattern: pattern }, false, 'setEupLoadingPattern'),
+    setIsEUPStackable: (stackable) => set({ isEUPStackable: stackable }, false, 'setIsEUPStackable'),
+    setIsDINStackable: (stackable) => set({ isDINStackable: stackable }, false, 'setIsDINStackable'),
+    setStackingStrategy: (strategy) => set({ stackingStrategy: strategy }, false, 'setStackingStrategy'),
+    setLastEdited: (last) => set({ lastEdited: last }, false, 'setLastEdited'),
+
+    clearAllPallets: () =>
+      set(
+        () => {
+          const emptyEntries = createEmptyWeightEntries();
+          const truckConfig = TRUCK_TYPES[get().selectedTruck];
+          const emptyArrangement = truckConfig.units.map((unit) => ({
+            unitId: unit.id,
+            unitLength: unit.length,
+            unitWidth: unit.width,
+            pallets: [],
+          }));
+
+          return {
+            ...emptyEntries,
+            isEUPStackable: false,
+            isDINStackable: false,
+            eupLoadingPattern: 'auto',
+            lastEdited: 'eup',
+            loadedEuroPalletsBase: 0,
+            loadedIndustrialPalletsBase: 0,
+            totalEuroPalletsVisual: 0,
+            totalDinPalletsVisual: 0,
+            utilizationPercentage: 0,
+            warnings: [],
+            palletArrangement: emptyArrangement,
+            totalWeightKg: 0,
+            actualEupLoadingPattern: 'auto',
+            remainingCapacity: { eup: 0, din: 0 },
+          };
+        },
+        false,
+        'clearAllPallets',
+      ),
+
+    maximizePallets: (palletTypeToMax) => {
+      const state = get();
+      const simEupWeights =
+        palletTypeToMax === 'euro'
+          ? [{ id: 1, quantity: MAX_PALLET_SIMULATION_QUANTITY, weight: '0', stackable: state.isEUPStackable }]
+          : [];
+      const simDinWeights =
+        palletTypeToMax === 'industrial'
+          ? [{ id: 1, quantity: MAX_PALLET_SIMULATION_QUANTITY, weight: '0', stackable: state.isDINStackable }]
+          : [];
+      const simResults = calculateLoadingLogic(
+        state.selectedTruck,
+        simEupWeights,
+        simDinWeights,
+        state.isEUPStackable,
+        state.isDINStackable,
+        'auto',
+        palletTypeToMax === 'euro' ? 'EUP_FIRST' : 'DIN_FIRST',
+        computeStackableCount(simEupWeights),
+        computeStackableCount(simDinWeights),
+        state.stackingStrategy,
+      );
+
+      if (palletTypeToMax === 'industrial') {
+        set(
+          () => ({
+            dinWeights: [{ id: Date.now(), weight: '', quantity: simResults.totalDinPalletsVisual, stackable: state.isDINStackable }],
+            eupWeights: [{ id: Date.now() + 1, weight: '', quantity: 0, stackable: state.isEUPStackable }],
+            lastEdited: 'din',
+          }),
+          false,
+          'maximizePallets',
+        );
+      } else {
+        set(
+          () => ({
+            eupWeights: [{ id: Date.now(), weight: '', quantity: simResults.totalEuroPalletsVisual, stackable: state.isEUPStackable }],
+            dinWeights: [{ id: Date.now() + 1, weight: '', quantity: 0, stackable: state.isDINStackable }],
+            lastEdited: 'eup',
+          }),
+          false,
+          'maximizePallets',
+        );
+      }
     },
 
-    reset: () => set({
-      units: 'metric',
-      containerId: CONTAINER_PRESETS[0]?.id ?? 'sprinter',
-      containerCustom: undefined,
-      palletId: PALLET_PRESETS[0]?.id ?? 'euro',
-      palletCustom: undefined,
-      constraints: defaultConstraints,
-      note: undefined,
-      plan: null,
-      lastValidationErrors: [],
-    }, false, 'reset'),
-  }))
+    fillRemaining: (typeToFill) => {
+      const state = get();
+      const weightEntryToUse = typeToFill === 'euro' ? state.eupWeights[state.eupWeights.length - 1] : state.dinWeights[state.dinWeights.length - 1];
+      const weightToFill = weightEntryToUse?.weight || '0';
+
+      const eupSim = typeToFill === 'euro'
+        ? [...state.eupWeights, { id: -1, quantity: MAX_PALLET_SIMULATION_QUANTITY, weight: weightToFill, stackable: state.isEUPStackable }]
+        : [...state.eupWeights];
+      const dinSim = typeToFill === 'industrial'
+        ? [...state.dinWeights, { id: -1, quantity: MAX_PALLET_SIMULATION_QUANTITY, weight: weightToFill, stackable: state.isDINStackable }]
+        : [...state.dinWeights];
+
+      const order = typeToFill === 'euro' ? 'DIN_FIRST' : 'EUP_FIRST';
+
+      const res = calculateLoadingLogic(
+        state.selectedTruck,
+        eupSim,
+        dinSim,
+        state.isEUPStackable,
+        state.isDINStackable,
+        'auto',
+        order,
+        computeStackableCount(eupSim),
+        computeStackableCount(dinSim),
+        state.stackingStrategy,
+      );
+
+      const currentEups = state.eupWeights.reduce((s, e) => s + e.quantity, 0);
+      const currentDins = state.dinWeights.reduce((s, e) => s + e.quantity, 0);
+
+      const addedEups = res.totalEuroPalletsVisual - currentEups;
+      const addedDins = res.totalDinPalletsVisual - currentDins;
+
+      let updated = false;
+
+      set(
+        (existing) => {
+          const updates: Partial<PlannerState> = { lastEdited: typeToFill === 'euro' ? 'eup' : 'din' };
+
+          if (typeToFill === 'euro' && addedEups > 0 && existing.eupWeights.length > 0) {
+            const newWeights = [...existing.eupWeights];
+            const lastIndex = newWeights.length - 1;
+            const updatedLastEntry = {
+              ...newWeights[lastIndex],
+              quantity: newWeights[lastIndex].quantity + addedEups,
+            };
+            newWeights[lastIndex] = updatedLastEntry;
+            updates.eupWeights = newWeights;
+            updated = true;
+          } else if (typeToFill === 'industrial' && addedDins > 0 && existing.dinWeights.length > 0) {
+            const newWeights = [...existing.dinWeights];
+            const lastIndex = newWeights.length - 1;
+            const updatedLastEntry = {
+              ...newWeights[lastIndex],
+              quantity: newWeights[lastIndex].quantity + addedDins,
+            };
+            newWeights[lastIndex] = updatedLastEntry;
+            updates.dinWeights = newWeights;
+            updated = true;
+          }
+
+          return updates;
+        },
+        false,
+        'fillRemaining',
+      );
+
+      return updated;
+    },
+
+    recalculate: () => {
+      const state = get();
+      const eupQuantity = state.eupWeights.reduce((sum, entry) => sum + entry.quantity, 0);
+      const dinQuantity = state.dinWeights.reduce((sum, entry) => sum + entry.quantity, 0);
+      const maxStackableEup = computeStackableCount(state.eupWeights);
+      const maxStackableDin = computeStackableCount(state.dinWeights);
+
+      const primaryResults = calculateLoadingLogic(
+        state.selectedTruck,
+        state.eupWeights,
+        state.dinWeights,
+        state.isEUPStackable,
+        state.isDINStackable,
+        state.eupLoadingPattern,
+        'DIN_FIRST',
+        maxStackableEup,
+        maxStackableDin,
+        state.stackingStrategy,
+      );
+
+      let multiTruckWarnings: string[] = [];
+
+      if (dinQuantity > 0 && eupQuantity === 0) {
+        const dinCapacitySim = [{ id: 1, quantity: MAX_PALLET_SIMULATION_QUANTITY, weight: '0', stackable: state.isDINStackable }];
+        const dinCapacityResult = calculateLoadingLogic(
+          state.selectedTruck,
+          [],
+          dinCapacitySim,
+          state.isEUPStackable,
+          state.isDINStackable,
+          state.eupLoadingPattern,
+          'DIN_FIRST',
+          maxStackableEup,
+          computeStackableCount(dinCapacitySim),
+          state.stackingStrategy,
+        );
+        const maxDinCapacity = dinCapacityResult.totalDinPalletsVisual;
+
+        if (maxDinCapacity > 0 && dinQuantity > maxDinCapacity) {
+          const totalTrucks = Math.ceil(dinQuantity / maxDinCapacity);
+          const fullTrucks = Math.floor(dinQuantity / maxDinCapacity);
+          const remainingPallets = dinQuantity % maxDinCapacity;
+
+          if (remainingPallets === 0) {
+            multiTruckWarnings.push(`Für diesen Auftrag werden ${fullTrucks} volle LKWs benötigt.`);
+          } else {
+            multiTruckWarnings.push(`Benötigt ${totalTrucks} LKWs: ${fullTrucks} volle LKW(s) und 1 LKW mit ${remainingPallets} Paletten.`);
+          }
+        }
+      } else if (eupQuantity > 0 && dinQuantity === 0) {
+        const eupCapacitySim = [{ id: 1, quantity: MAX_PALLET_SIMULATION_QUANTITY, weight: '0', stackable: state.isEUPStackable }];
+        const eupCapacityResult = calculateLoadingLogic(
+          state.selectedTruck,
+          eupCapacitySim,
+          [],
+          state.isEUPStackable,
+          state.isDINStackable,
+          state.eupLoadingPattern,
+          'EUP_FIRST',
+          computeStackableCount(eupCapacitySim),
+          maxStackableDin,
+          state.stackingStrategy,
+        );
+        const maxEupCapacity = eupCapacityResult.totalEuroPalletsVisual;
+
+        if (maxEupCapacity > 0 && eupQuantity > maxEupCapacity) {
+          const totalTrucks = Math.ceil(eupQuantity / maxEupCapacity);
+          const fullTrucks = Math.floor(eupQuantity / maxEupCapacity);
+          const remainingPallets = eupQuantity % maxEupCapacity;
+
+          if (remainingPallets === 0) {
+            multiTruckWarnings.push(`Für diesen Auftrag werden ${fullTrucks} volle LKWs benötigt.`);
+          } else {
+            multiTruckWarnings.push(`Benötigt ${totalTrucks} LKWs: ${fullTrucks} volle LKW(s) und 1 LKW mit ${remainingPallets} Paletten.`);
+          }
+        }
+      }
+
+      const weightToFillEup = state.eupWeights.length > 0 ? state.eupWeights[state.eupWeights.length - 1].weight || '0' : '0';
+      const eupCapacitySim = [{ id: -1, quantity: MAX_PALLET_SIMULATION_QUANTITY, weight: weightToFillEup, stackable: state.isEUPStackable }];
+      const eupCapacityResult = calculateLoadingLogic(
+        state.selectedTruck,
+        eupCapacitySim,
+        state.dinWeights,
+        state.isEUPStackable,
+        state.isDINStackable,
+        state.eupLoadingPattern,
+        'DIN_FIRST',
+        computeStackableCount(eupCapacitySim),
+        maxStackableDin,
+        state.stackingStrategy,
+      );
+      const maxEup = eupCapacityResult.totalEuroPalletsVisual;
+      const remainingEup = Math.max(0, maxEup - eupQuantity);
+
+      const weightToFillDin = state.dinWeights.length > 0 ? state.dinWeights[state.dinWeights.length - 1].weight || '0' : '0';
+      const dinCapacitySim = [{ id: -1, quantity: MAX_PALLET_SIMULATION_QUANTITY, weight: weightToFillDin, stackable: state.isDINStackable }];
+      const dinCapacityResult = calculateLoadingLogic(
+        state.selectedTruck,
+        state.eupWeights,
+        dinCapacitySim,
+        state.isEUPStackable,
+        state.isDINStackable,
+        state.eupLoadingPattern,
+        'EUP_FIRST',
+        maxStackableEup,
+        computeStackableCount(dinCapacitySim),
+        state.stackingStrategy,
+      );
+      const maxDin = dinCapacityResult.totalDinPalletsVisual;
+      const remainingDin = Math.max(0, maxDin - dinQuantity);
+
+      set(
+        {
+          palletArrangement: primaryResults.palletArrangement,
+          loadedIndustrialPalletsBase: primaryResults.loadedIndustrialPalletsBase,
+          loadedEuroPalletsBase: primaryResults.loadedEuroPalletsBase,
+          totalDinPalletsVisual: primaryResults.totalDinPalletsVisual,
+          totalEuroPalletsVisual: primaryResults.totalEuroPalletsVisual,
+          utilizationPercentage: primaryResults.utilizationPercentage,
+          warnings: Array.from(new Set([...primaryResults.warnings, ...multiTruckWarnings])),
+          totalWeightKg: primaryResults.totalWeightKg,
+          actualEupLoadingPattern: primaryResults.eupLoadingPatternUsed,
+          remainingCapacity: { eup: remainingEup, din: remainingDin },
+        },
+        false,
+        'recalculate',
+      );
+    },
+  })),
 );
